Use user:pass@ from URI in Authorization: Basic
simpleHTTP allows the URI of the format "http://user:pass@host.dom/path",
but user:pass@ part is simply stripped out and ignored. This commit
introduces a 'normalizer' that takes user:pass@ data from the request's
URI and adds "Authorization: Basic ..." header to the request. This
normalizer is called before the normalizeHostURI normalizer, as the
latter strips the user:pass@ part from the URI.