fix openTCPConnection on Windows.
The problem was that getAddrInfo doesn't initialize the socket library
implicitly. If that is the first thing that is done with sockets, then
the program will crash and burn with a misleading error message.

The initialization is done only once, so the cost of a superfluous call is
minimal. It is a no-op on other OSes.