Replace derived Eq instance for HeaderName with manual instance (fixes #128)
This change replaces the derived Eq instance with one that causes custom
headers to be compared case-insensitively. This is because situations
where header name matching is done (such as via lookupHeader along with
many other cases) should ignore case, but prior to this change it was
possible to mismatch headers even when they had the same
(case-insensitive) name.